# Fix performance regression in ConvexPolygonTestPerformance 