--- a/Framework/DataObjects/inc/MantidDataObjects/CalculateReflectometry.h
+++ b/Framework/DataObjects/inc/MantidDataObjects/CalculateReflectometry.h
@@ -13,13 +13,9 @@ namespace DataObjects {
 class CalculateReflectometry {
 
 protected:
-  const double to_radians_factor;
+  constexpr double to_radians_factor{M_PI / 180.0};
 
 public:
-  /**
-   Constructor
-   */
-  CalculateReflectometry() : to_radians_factor(M_PI / 180.0) {}
 
   virtual ~CalculateReflectometry() = default;