--- a/Framework/DataObjects/inc/MantidDataObjects/CalculateReflectometry.h
+++ b/Framework/DataObjects/inc/MantidDataObjects/CalculateReflectometry.h
@@ -16,7 +16,6 @@ protected:
   constexpr double to_radians_factor{M_PI / 180.0};
 
 public:
-
   virtual ~CalculateReflectometry() = default;
 
   /**