--- a/Framework/Geometry/test/ConvexPolygonTest.h
+++ b/Framework/Geometry/test/ConvexPolygonTest.h
@@ -71,16 +71,15 @@ public:
 
   void test_Area_Of_A_Parallelogram() {
     ConvexPolygon parallelogram = makeParallelogram();
-    TS_ASSERT_DELTA(parallelogram.area(), std::sqrt(2.0), DBL_EPSILON);
+    TS_ASSERT_DELTA(parallelogram.area(), M_SQRT2, DBL_EPSILON);
   }
 
   void test_Extreme_Points_Are_Correct() {
     ConvexPolygon parallelogram = makeParallelogram();
     TS_ASSERT_DELTA(parallelogram.minX(), 0.0, DBL_EPSILON);
-    TS_ASSERT_DELTA(parallelogram.maxX(), 2.0 + 0.5 * std::sqrt(2.0),
-                    DBL_EPSILON);
+    TS_ASSERT_DELTA(parallelogram.maxX(), 2.0 + 0.5 * M_SQRT2, DBL_EPSILON);
     TS_ASSERT_DELTA(parallelogram.minY(), 0.0, DBL_EPSILON);
-    TS_ASSERT_DELTA(parallelogram.maxY(), 0.5 * std::sqrt(2.0), DBL_EPSILON);
+    TS_ASSERT_DELTA(parallelogram.maxY(), 0.5 * M_SQRT2, DBL_EPSILON);
   }
 
   void test_Polygon_Contains_Polygon() {
@@ -147,8 +146,8 @@ private:
   ConvexPolygon makeParallelogram() {
     ConvexPolygon parallelogram;
     parallelogram.insert(0.0, 0.0);
-    parallelogram.insert(0.5 * std::sqrt(2.0), 0.5 * std::sqrt(2.0));
-    parallelogram.insert(2.0 + 0.5 * std::sqrt(2.0), 0.5 * std::sqrt(2.0));
+    parallelogram.insert(0.5 * M_SQRT2, 0.5 * M_SQRT2);
+    parallelogram.insert(2.0 + 0.5 * M_SQRT2, 0.5 * M_SQRT2);
     parallelogram.insert(2.0, 0.0);
     return parallelogram;
   }
@@ -159,17 +158,29 @@ private:
 //------------------------------------------------------------------------------
 class ConvexPolygonTestPerformance : public CxxTest::TestSuite {
 public:
-  void test_Area_Calls() {
-    const size_t ntests(50000000);
-
-    double totalArea(0.0);
+  // FractionalRebinning constructs ConvexPolygon once, then updates its values.
+  void test_update_values() {
+    const size_t ntests{1000000};
+    ConvexPolygon test;
     for (size_t i = 0; i < ntests; ++i) {
-      ConvexPolygon test;
+      test.clear();
       test.insert(0.0, 0.0);
       test.insert(0.0, 1.0);
       test.insert(2.0, 1.0);
       test.insert(2.0, 0.0);
+    }
+  }
 
+  void test_Area_Calls() {
+    ConvexPolygon test;
+    test.insert(0.0, 0.0);
+    test.insert(0.0, 1.0);
+    test.insert(2.0, 1.0);
+    test.insert(2.0, 0.0);
+
+    const size_t ntests{10000000};
+    double totalArea{0.0};
+    for (size_t i = 0; i < ntests; ++i) {
       totalArea += test.area();
     }
   }