Description of work.

This addresses the `ConvexPolygonlTestPerformance` and `QuadrilateralTestPerformance` regressions discussed in #15856. with GCC on linux, switching Vertices from a `std::deque` to a `std::vector` eliminates calls to `malloc_consolidate` and improves performance in the old test by about an 10x. 

I split up the performance tests into two, one that updates the points in a `ConvexPolygon` and another that tests calculating the area of the polyhedron. 

**To test:**

<!-- Instructions for testing. -->

This PR is best verified on Linux since the performance regression occurs there. Since I modified the performance tests, use the test currently in master to verify that the performance regression is fixed. 

---

#### Reviewer ####

Please comment on the following ([full description](http://www.mantidproject.org/Individual_Ticket_Testing)):

##### Code Review #####

- [x] Is the code of an acceptable quality?
- [x] Does the code conform to the coding standards? Is it well structured with small focussed classes/methods/functions?
- [x] Are there unit/system tests in place? Are the unit tests small and test the a class in isolation?
- [x] If there are changes in the release notes then do they describe the changes appropriately?

##### Functional Tests #####

- [x] Do changes function as described? Add comments below that describe the tests performed?
- [x] How do the changes handle unexpected situations, e.g. bad input?
- [x] Has the relevant documentation been added/updated?
- [x] Is user-facing documentation written in a user-friendly manner?
- [x] Has developer documentation been updated if required?

- Does everything look good? Comment with the ship it emoji but don't merge. A member of `@mantidproject/gatekeepers` will take care of it.