@MikeHart85 2016-05-11
Tested on Linux. 

Significant improvement to both the `QuadrilateralTestPerformance` and the `ConvexPolygonTestPerformance` tests. Using the old tests, runtime on my machine went from 6.18 sec to 0.57 sec and from 9.5 sec to 7.4 sec, respectively, as compared to #15856. The new tests split testing construction performance from area calculation performance, and `ConvexPolygonTestPerformance` runs much faster since it uses fewer iterations.

:shipit: